Patch from Sebastian Audet
KMines Reset After Game Over.

At the end of the game you are prompted to restart the same or continue to a new one.

You have the option of saying 'no' - besides that unless you're in a competition I don't think it should matter. Mainly I proposed the change because there are times when its impossible to do anything but guess (not even statistics) and this is a midway between a 'reveal' button and just flat out spending 4 minutes on something and failing.

Differential Revision: https://phabricator.kde.org/D17860